requirements.txt
aiosmtpd==1.4.6
atpublic==7.0.0
attrs==25.4.0
azure-core==1.37.0
azure-data-tables==12.7.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.45.1
yarl==1.22.0
